+++ semcore/notice/src/index.d.ts
@@ -3,6 +3,7 @@ import { Box, BoxProps } from '@semcore/flex-box';
 import Button from '@semcore/button';
 import { IconProps } from '@semcore/icon';
 import { FadeInOutProps } from '@semcore/animation';
+import { Text } from '@semcore/typography';
 
 export type NoticeTheme = 'danger' | 'warning' | 'success' | 'info' | 'muted';
 export type NoticeUse = 'primary' | 'secondary';
@@ -72,8 +73,8 @@ declare const Notice: Intergalactic.Component<'div', NoticeProps, NoticeContext>
   Label: Intergalactic.Component<'div', NoticeLabelProps>;
   Actions: typeof Box;
   Content: typeof Box;
-  Title: typeof Box;
-  Text: typeof Box;
+  Title: typeof Text;
+  Text: typeof Text;
   /**
    * @deprecated Use Notice.Close instead of Notice.CloseIcon
    */